backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
python-multipart==0.0.12
pytesseract==0.3.13
pillow==11.0.0
pdfplumber==0.11.4
PyPDF2==3.0.1
reportlab==4.2.5
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
groq==1.0.0
playwright==1.49.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
